fixing query for selecting instruments

+++ src/database/asset.rs
@@ -17,12 +17,12 @@ impl Assets {
     /// * `handler` - &Handler
     pub async fn get_assets(
         handler: &Handler,
-    ) -> Result<HashMap<String, i64>, Box<dyn std::error::Error>> {
+    ) -> Result<HashMap<String, i32>, Box<dyn std::error::Error>> {
         let assets = sqlx::query_as::<_, Assets>(r#"SELECT "Id", "Code" FROM "Assets""#)
             .fetch_all(&handler.pool)
             .await?;
 
-        let assets_map: HashMap<String, i64> = assets.into_iter().map(|a| (a.code, a.id as i64)).collect();
+        let assets_map: HashMap<String, i32> = assets.into_iter().map(|a| (a.code, a.id)).collect();
 
         Ok(assets_map)
     }
