+++ src/sources/mod.rs
@@ -11,7 +11,7 @@ pub mod rest_source;
 pub trait SourceOps {
     fn fetch(
         &self,
-        db_assets: HashMap<String, i64>,
+        db_assets: HashMap<String, i32>,
         db_instruments: HashMap<String, Instrument>,
         opts: &Opts,
     ) -> Result<Vec<(DBInstrument, String)>, Box<dyn std::error::Error>>;
